Splitting up hearing prep feature tests
